+++ docs/source/release_notes.rst
@@ -14,6 +14,7 @@ Version 1.0
 - Remove previously deprecated functions
 - Add :class:`cellrank.ul.models.FailedModel` inspired by the maybe monad
 - Allow returning models when doing bulk fitting
+- Add ``transpose`` parameter for :func:`cellrank.pl.gene_trends`
 - Various other minor bugfixes
 
 1.0.0-rc.11 :small:`2020-09-25`